fix type discrepancies on windows

+++ _unittests/ut_array_api/test_onnx_numpy.py
@@ -90,7 +90,7 @@ class TestOnnxNumpy(ExtTestCase):
         mat = xp.arange(a, b)
         matnp = mat.numpy()
         self.assertEqual(matnp.shape, (0,))
-        expected = np.arange(0, 0)
+        expected = np.arange(0, 0).astype(np.int64)
         self.assertEqualArray(matnp, expected)
 
     @ignore_warnings(DeprecationWarning)
@@ -98,7 +98,7 @@ class TestOnnxNumpy(ExtTestCase):
         mat = xp.arange(0, 0)
         matnp = mat.numpy()
         self.assertEqual(matnp.shape, (0,))
-        expected = np.arange(0, 0)
+        expected = np.arange(0, 0).astype(np.int64)
         self.assertEqualArray(matnp, expected)
 
     def test_ones_like_uint16(self):
